refactor(checker): split 4 over-cap checker files post #12101 steps 1-7
Pure behavior-preserving file split to bring the four checker source files
that the combined #12101 on-demand-forcing steps 1-7 pushed over the 2000-LOC
architecture cap (and constructor.rs over its 1981 ratchet) back under their
limits. No forcing logic changed; the byte-parity-validated #12101 path is
untouched. Only code moves (verbatim extraction) and comment trims.

Extractions (all verbatim, methods stay on `CheckerState`):
- assignability_checker.rs 2035 -> 1267: the `should_suppress_*` diagnostic
  suppression cluster moved to `query_boundaries/assignability_suppression.rs`
  (also drops 63 `query_boundaries::common` direct refs out of the #8225 count
  since the new home is inside `query_boundaries/`, so no cap bump is needed).
- class.rs 2014 -> 1615: the class-decorator checks moved to
  `state/state_checking/class_decorators.rs`; the four helpers reached from the
  class-declaration/expression paths were widened `fn` -> `pub(super)` solely
  to permit the split.
- constructor.rs 1987 -> 1976 (ratchet 1981): the `StaticMemberBuildData` input
  struct moved to `constructor_parts/build_data.rs` (`pub(super)` fields);
  `helpers.rs` updated to import it from the new module.
- context/mod.rs 2001 -> 1999: trimmed an over-verbose #12101 RefCell doc block.

Arch-contract tests that pin routing strings by file path were updated to read
the new module homes (keyof suppression -> query_boundaries module; class
decorator return routing -> class_decorators.rs); behavior assertions unchanged.

Verification: arch guard exit 0 (all four files + the constructor ratchet pass,
#8225 and CheckerContext field caps unchanged); cargo build/clippy clean, fmt
clean; full `cargo test -p tsz-checker --lib` failing set identical to the
pre-refactor head excluding the known `source_option_bag` flake (empty delta).

Refs #12101.
